requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
praw==7.6.1
pandas==1.5.0
bertopic==0.15.0
sentence-transformers==2.2.2
ollama==0.0.14
python-dotenv==1.0.0 